pyproject.toml
[tool.poetry]
name = "worklife-test"
version = "0.3.0"
description = ""
authors = []


[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
SQLAlchemy = "^2.0"
pydantic = "^2.7"
psycopg2-binary = "^2.9.3"
alembic = "^1.9.4"
fastapi-utils = "^0.7"
typing-inspect = "^0.9.0"
pydantic-settings = "^2.4.0"
factory-boy = "^3.3.1"

[tool.poetry.dev-dependencies]
black = "^24"
pytest = "^8.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14"
